third_party/WebKit/LayoutTests/fast/events/autoscroll-over-scrollbar.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/compositor-controls.js"></script>
<style>
  body {
    margin: 0px;
    height: 1000px;
    width: 1000px;
  }
  #scrollable {
    background-color: #FF7F7F;
    height: 600px;
    width: 600px;
    overflow: scroll;
  }
  #content {
    height: 700px;
    width: 700px;
  }
</style>

<div id="scrollable">
  <div id="content"> </div>
</div>

<script>
setAnimationRequiresRaster();
var scrollable = document.getElementById('scrollable');
function autoScroll(start_x, start_y, end_x, end_y) {
  return new Promise((resolve, reject) => {
    if (!window.eventSender) {
      reject();
    } else {
      const MIDDLE_BUTTON = 1;
      eventSender.mouseMoveTo(start_x, start_y);
      eventSender.mouseDown(MIDDLE_BUTTON);
      eventSender.mouseUp(MIDDLE_BUTTON);
      eventSender.mouseMoveTo(end_x, end_y);
      resolve();
    }
  });
}

function waitForScrollAndCheck() {
  const MAX_FRAME = 500;
  return new Promise((resolve, reject) => {
    function tick(frames) {
    // We requestAnimationFrame either for 500 frames or until scrollable scrolls.
      if (frames >= MAX_FRAME || scrollable.scrollTop > 0)
        resolve();
      else
        requestAnimationFrame(tick.bind(this, frames + 1));
    }
    tick(0);
  });
}

promise_test(t => {
  const MIDDLE_CLICK_AUTOSCROLL_RADIUS = 15; // from blink::kNoMiddleClickAutoscrollRadius
  var rect = scrollable.getBoundingClientRect();
  var startX = rect.right - 5;
  var startY = rect.top + 20;
  var endX = startX;
  var endY = startY + 2 * MIDDLE_CLICK_AUTOSCROLL_RADIUS;
  assert_equals(scrollable.scrollTop, 0);
  return autoScroll(startX, startY, endX, endY)
  .then(waitForScrollAndCheck)
  .then(() => {
      assert_greater_than(scrollable.scrollTop, 0);
  });
});

</script>
